+++ dao/dao.GetObject_test.go
@@ -16,27 +16,28 @@ func TestDAOGetObject(t *testing.T) {
 	var obj models.ODObject
 	obj.Name = "Test Object for GetObject"
 	obj.CreatedBy = usernames[1]
-	obj.TypeName.String = "File"
-	obj.TypeName.Valid = true
+	obj.TypeName = models.ToNullString("File")
+
 	permissions := make([]models.ODObjectPermission, 1)
 	permissions[0].Grantee = obj.CreatedBy
 	permissions[0].AcmShare = fmt.Sprintf(`{"users":[%s]}`, permissions[0].Grantee)
 	permissions[0].AcmGrantee.Grantee = permissions[0].Grantee
-	permissions[0].AcmGrantee.UserDistinguishedName.String = permissions[0].Grantee
-	permissions[0].AcmGrantee.UserDistinguishedName.Valid = true
+	permissions[0].AcmGrantee.UserDistinguishedName = models.ToNullString(permissions[0].Grantee)
 	permissions[0].AllowCreate = true
 	permissions[0].AllowRead = true
 	permissions[0].AllowUpdate = true
 	permissions[0].AllowDelete = true
 	obj.Permissions = permissions
+
 	properties := make([]models.ODObjectPropertyEx, 1)
 	properties[0].Name = "Test Property in TestDAOGetObject"
-	properties[0].Value.String = "Test Property Value"
-	properties[0].Value.Valid = true
-	properties[0].ClassificationPM.String = "UNCLASSIFIED"
-	properties[0].ClassificationPM.Valid = true
+	properties[0].Value = models.ToNullString("Test Property Value")
+
+	properties[0].ClassificationPM = models.ToNullString("UNCLASSIFIED")
 	obj.Properties = properties
-	obj.RawAcm.String = testhelpers.ValidACMUnclassified
+
+	obj.RawAcm = models.ToNullString(testhelpers.ValidACMUnclassified)
+
 	dbObject, err := d.CreateObject(&obj)
 	if err != nil {
 		t.Error(err)
@@ -51,30 +52,12 @@ func TestDAOGetObject(t *testing.T) {
 		t.Error("expected TypeID to be set")
 	}
 
-	// // add property
-	// var property models.ODProperty
-	// property.CreatedBy = dbObject.CreatedBy
-	// property.Name = "Test Property"
-	// property.Value.String = "Test Property Value"
-	// property.Value.Valid = true
-	// property.ClassificationPM.String = "UNCLASSIFIED"
-	// property.ClassificationPM.Valid = true
-	// _, err = d.AddPropertyToObject(dbObject, &property)
-	// if err != nil {
-	// 	t.Error(err)
-	// }
-
 	// get object with properties
 	objectWithProperty, err := d.GetObject(dbObject, true)
 	if err != nil {
 		t.Error(err)
 	}
 	if len(objectWithProperty.Properties) != 1 {
-
-		// jsonData, _ := json.MarshalIndent(objectWithProperty, "", "  ")
-		// jsonified := string(jsonData)
-		// fmt.Println(jsonified)
-
 		t.Errorf("Expected one property on the object, got %d", len(objectWithProperty.Properties))
 	} else {
 		if objectWithProperty.Properties[0].Name != "Test Property in TestDAOGetObject" {